api/requirements.txt
azure-search-documents==11.6.0b4
azure-identity
fastapi
uvicorn
python-dotenv==1.0.0
tiktoken
jira
langchain-openai==0.1.1
langchain==0.1.14
langchain-community==0.0.31
langchain-mongodb==0.1.3